+++ CHANGELOG.md
@@ -25,6 +25,7 @@ Helm release
 
 -   Support replicas in SDM goal data
 -   Helm charts
+-   Support for TLS secrets in ingress
 
 ## [0.7.3][] - 2018-05-01